requirements.txt
kivy==2.3.1
requests==2.25.1
python-dotenv==1.0.1
